added user verification of deletion

+++ manager.py
@@ -2,6 +2,7 @@
 
 
 import argparse
+import input
 import logging
 import sys
 
@@ -118,6 +119,18 @@ if args.mode == 'repo':
             sys.exit(-1)
 
     elif args.cmd == 'delete':
+        yesno = input('Are you sure you want to delete the repository "%s"?  This cannot be undone. [yes/NO] ' % args.name)
+
+        if yesno != 'yes':
+            print 'Repository will not be deleted'
+            sys.exit(0)
+
+        yesno = input('Verify the name of the repository: ')
+
+        if yesno != args.name:
+            print 'Repository will not be deleted'
+            sys.exit(0)
+
         logging.info('deleting repository %s', args.name)
 
         if not r.delete():
